% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_functions.R
\name{cpm_to_counts}
\alias{cpm_to_counts}
\title{Convert CPM Values Back to Counts}
\usage{
cpm_to_counts(data, library_size)
}
\arguments{
\item{data}{a matrix or matrix-like object where rows are genes and columns
are samples. All values in \code{data} should be >= 0.}

\item{library_size}{a numeric vector describing the library size of each
sample in \code{data}. The length of \code{library_size} must match the number of
columns in \code{data}. Typically library size is the sum of all counts in a
sample, but this could also be modified with TMM or other size factors as
well.}
}
\value{
an object the same type and shape as \code{data} where CPM values have
been converted back to integer counts. Any non-integer values will be
rounded off.
}
\description{
Convert CPM Values Back to Counts
}
\examples{
# Make a matrix of CPM data
counts <- round(matrix(runif(1000, min = 0, max = 1e8), nrow = 100))
cpm_data <- simple_cpm(counts)

# Reverse the operation
counts2 <- cpm_to_counts(cpm_data, library_size = colSums(counts))
all(counts == counts2)
}
\seealso{
\code{\link[=simple_cpm]{simple_cpm()}}
}
