% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{.get_removals}
\alias{.get_removals}
\title{Get Which Variables to Remove Based on Correlation}
\usage{
.get_removals(
  cor_mat,
  na_vars,
  R2_threshold = 0.5,
  removed = c(),
  always_keep = c()
)
}
\arguments{
\item{cor_mat}{a matrix of correlation (or other association-like) values,
which must have row and column names. This matrix does not need to be
square.}

\item{na_vars}{a one-row data.frame where the columns are the covariates and
the values are the number of \code{NA} values in each column.}

\item{R2_threshold}{(optional) the threshold for R^2, where variables with an
R^2 value > \code{R2_threshold} are considered for removal. Defaults to 0.5.}

\item{removed}{(optional) a string or character vector of variables that have
already been marked for removal. In the case where one covariate in a pair
of highly-correlated variables is in \code{removed} already, the other variable
will not be removed even if it would otherwise have met the criteria for
removal. Defaults to an empty vector.}

\item{always_keep}{(optional) a string or character vector of columns that
should always be kept even if they are highly correlated with another
variable. In that case, the other variable is removed instead, as long as
it is not also in \code{always_keep}. Defaults to an empty vector.}
}
\value{
a character vector with the names of the columns that should be
removed. May also be an empty vector.
}
\description{
This is a helper function for \code{\link[=remove_correlated_covariates]{remove_correlated_covariates()}} to decide
which variable in each highly-correlated pair of covariates should get
removed.
}
\examples{
\dontrun{
}
}
