% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{find_pca_outliers_by_group}
\alias{find_pca_outliers_by_group}
\title{Find Outliers by PCA, Split by Group}
\usage{
find_pca_outliers_by_group(
  data,
  pca_group,
  n_sds = 4,
  metadata = NULL,
  sample_colname = "specimenID",
  gene_info = NULL
)
}
\arguments{
\item{data}{a matrix, data.frame, or matrix-like object of expression data
where rows are genes and columns are samples. Data should be normalized and
on the log or log2 scale, for example as returned by \code{\link[=simple_lognorm]{simple_lognorm()}}.}

\item{pca_group}{either a vector or a single character string describing the
groups of samples. If \code{metadata} is provided, \code{pca_group} should be the
name of the column in \code{metadata} that provides the group values. If
\code{metadata} is missing or \code{NULL}, \code{pca_group} should be a vector the same
length as the number of columns in \code{data}, where each entry is the group
assignment of the corresponding sample in \code{data}.}

\item{n_sds}{(optional) samples will be labeled as outliers if they are
outside the ellipse defined by \code{radius1 = n_sds * sd(PC1)} and \code{radius2 = n_sds * sd(PC2)}. Defaults to 4.}

\item{metadata}{(optional) a data.frame where rows are samples and columns
are covariates. If supplied, it must contain, at minimum, a column with
sample IDs, and any additional columns will be merged into the result to
make plotting simpler. If omitted or \code{NULL}, the result will contain only
the columns returned in the \verb{$x} matrix from \code{prcomp}. See details for more
information.}

\item{sample_colname}{(optional) if \code{metadata} is provided, a character
string naming the column in \code{metadata} that contains sample labels. Ignored
if \code{metadata} is \code{NULL}. Defaults to "specimenID".}

\item{gene_info}{(optional) a data.frame where rows are genes and columns are
variables that provide metadata about each gene. If provided, the PCA will
run using only protein-coding autosomal genes, which are extracted from
\code{gene_info}. If \code{NULL} or not provided, the PCA will use all genes in
\code{data}. Defaults to \code{NULL}. See details for more information.}
}
\value{
a named list with the following items:

\item{group_results}{a list containing the return values from \code{\link[=find_pca_outliers]{find_pca_outliers()}} for each group}
\item{outliers}{a character vector of sample names that were marked as outliers in any group}

Each item in \code{group_results} can be used to plot the results with \code{\link[=plot_pca_outliers]{plot_pca_outliers()}}.
}
\description{
Finds outliers via PCA, but splits the data into groups and runs PCA outlier
detection on each group separately.
}
\details{
This function splits the input expression matrix into groups of samples as
specified by \code{pca_group} and runs PCA outlier detection on each group
separately. This is recommended if the data has noticeable batch effects or
comes from separate tissues. See \link{find_pca_outliers} for more details about
input, filtering, and plotting.
}
\examples{
\dontrun{
test
}
}
\seealso{
\code{\link[=find_pca_outliers]{find_pca_outliers()}}, \code{\link[=simple_lognorm]{simple_lognorm()}},
\code{\link[=plot_pca_outliers]{plot_pca_outliers()}}, \code{\link[=get_gc_content_biomart]{get_gc_content_biomart()}}, \code{\link[=get_gc_content_gtf]{get_gc_content_gtf()}}
}
