% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{find_sex_mismatches}
\alias{find_sex_mismatches}
\title{Find Mismatches Between Reported Sex and Y Chromosome Expression}
\usage{
find_sex_mismatches(
  metadata,
  data,
  sample_colname = "specimenID",
  sex_colname = "sex",
  y_expr_threshold = 2
)
}
\arguments{
\item{metadata}{a data.frame containing at least two columns, one for sample
IDs and one for the reported sex of each sample. The reported sex should be
either "male" or "female", all lowercase.}

\item{data}{a matrix, data.frame, or matrix-like object of expression data
where rows are genes and columns are samples. The rownames of data should
either be gene symbols or Ensembl IDs. Data should be normalized and on the
log or log2 scale, for example as returned by \code{\link[=simple_lognorm]{simple_lognorm()}}.}

\item{sample_colname}{(optional) a character string naming the column in
\code{metadata} that contains sample labels. Defaults to "specimenID".}

\item{sex_colname}{(optional) a character string naming the column in
\code{metadata} that contains sex labels. Defaults to "sex".}

\item{y_expr_threshold}{(optional) a numeric threshold that splits samples
based on mean expression of the Y-chromosome marker genes. Samples with
mean expression above the threshold will be labeled as "male" while samples
with mean expression below the threshold will be labeled as "female".
Defaults to 2.0, though this value may need to be be calibrated to the
dataset.}
}
\value{
a named list with the following items:

\item{sex_check_df}{a data.frame containing the reported sex, estimated sex,
and expression values of the sex marker genes for each sample. It also
contains a \code{sex_valid} column denoting whether the reported sex matches the
estimated sex}
\item{mismatches}{a character vector of samples whose estimated sex does not
match the reported sex}

The \code{sex_check_df} data frame can be used with \code{\link[=plot_sex_mismatch_results]{plot_sex_mismatch_results()}}
to view the results.
}
\description{
Estimates the sex of each sample based on expression of several genes on the
Y chromosome, and compares estimated sex to the reported sex to find
potential mismatches.
}
\details{
This function uses the mean expression of four Y-chromosome marker genes to
estimate the sex of each sample. The four genes are RPS4Y1, EIF1AY, DDX3Y,
and KDM5D, which were determined to be good markers for sex in \href{https://www.ncbi.nlm.nih.gov/pubmed/23829492}{Staedtler, et. al (2013)}.

\strong{Note:} This function currently only works for human genes.
}
\examples{
\dontrun{
test = 1
}
}
\seealso{
\code{\link[=plot_sex_mismatch_results]{plot_sex_mismatch_results()}}, \code{\link[=simple_lognorm]{simple_lognorm()}}
}
