% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{generalized_correlation}
\alias{generalized_correlation}
\title{Generalized correlation}
\usage{
generalized_correlation(data, exclude_cols = NULL)
}
\arguments{
\item{data}{a data.frame, matrix, or object coercible to a data.frame, where
rows are observations/samples and columns are variables.}

\item{exclude_cols}{(optional) a string or character vector containing the
names of any columns in \code{data} to exclude from this calculation, for
example an ID column. Defaults to \code{NULL}.}
}
\value{
an NxN symmetric matrix, where N = the number of columns of \code{data}
minus any \code{exclude_cols}. Values will be the correlation or Cramer's V
between each variable. Note that values for categorical vs categorical and
categorical vs numerical will always be positive, but numeric vs numeric
correlation may be positive or negative.
}
\description{
Calculates correlation or correlation-like values between numeric variables,
categorical variables, and combinations of both. Categorical vs categorical
is calculated with Cramer's V, and categorical vs numeric is calculated as
the square root of the R^2 value from a linear model of \code{numeric ~ categorical}.
}
\examples{
df <- data.frame(var1 = rnorm(10),
                 var2 = c(rnorm(5, mean = 0), rnorm(5, mean = 10)), # Correlated with group
                 var3 = c(rep("group1", 5), rep("group2", 5)),
                 var4 = c(rep("dose1", 3), rep("dose2", 3), rep("dose3", 4)))
generalized_correlation(df)
}
