% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_content_functions.R
\name{get_gc_content_biomart}
\alias{get_gc_content_biomart}
\title{Get Gene GC Content from BioMart Data}
\usage{
get_gc_content_biomart(
  gene_ids,
  include_introns = FALSE,
  mart = NULL,
  dataset = "hsapiens_gene_ensembl",
  ensembl_version = NULL,
  additional_attributes = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{gene_ids}{a character vector of Ensembl gene IDs to get GC content for.}

\item{include_introns}{(optional) if \code{FALSE}, GC content will be calculated
using exon sequences only. If \code{TRUE}, content will be calculated using the
entire gene sequence, including introns. Defaults to \code{FALSE}.}

\item{mart}{(optional) a \code{\link[biomaRt:Mart-class]{biomaRt::Mart}} object that
is connected to a specific BioMart database and dataset. If \code{NULL}, this
function will connect to BioMart using the \code{dataset} and \code{version}
arguments instead.}

\item{dataset}{(optional) the name of the dataset to use from BioMart, if
\code{mart} is \code{NULL}. Defaults to "hsapiens_gene_ensembl" for human genes. Both
"hsapiens_gene_ensembl" and "mmusculus_gene_ensembl" have been tested with
this function, but other species may work as well.}

\item{ensembl_version}{(optional) the Ensembl version to use (e.g. 110 or
113), if \code{mart} is \code{NULL} and data should be fetched from an older archive.
If \code{NULL}, the latest version of Ensembl will be queried.}

\item{additional_attributes}{(optional) a character vector of attributes /
columns to fetch from BioMart in addition to the default columns returned
by this function (see details). If \code{NULL}, only the default columns will be
used.}

\item{verbose}{(optional) whether to print status messages}

\item{...}{(optional) additional arguments to \code{\link[biomaRt:useEnsembl]{biomaRt::useEnsembl()}}, for
example an Ensembl mirror could be specified if BioMart is having
connection issues.}
}
\value{
a data.frame where rows are genes and columns are variables. It will
contain, at minimum, columns for \code{ensembl_gene_id}, \code{external_gene_name},
\code{gene_biotype}, \code{chromosome_name}, \code{gene_length}, and \code{percent_gc_content}.
It will also contain columns listed in the \code{additional_attributes}
argument.
}
\description{
Queries BioMart for exon coordinates and sequences for each gene and
calculates GC content from the results.
}
\details{
This function is a wrapper around \code{\link[=calculate_gc_content]{calculate_gc_content()}}, which calculates
GC content based on either the entire gene, including both introns and exons,
or just exon sequences only. Coordinates and sequences are obtained from
BioMart, which may or may not have data on all genes present in the counts
matrix.

Previous versions of Ensembl can be used instead of the latest version by
adding \verb{version = <Ensembl version>} as an argument. GC content will be the
most accurate if the Ensembl version corresponding to the alignment reference
is used.

This function obtains the following information from BioMart for each gene:
\code{ensembl_gene_id}, \code{ensembl_exon_id}, \code{exon_chrom_start}, \code{exon_chrom_end},
\code{strand}, \code{start_position}, \code{end_position}, \code{external_gene_name},
\code{chromosome_name}, \code{gene_biotype}.

To add additional columns from Biomart, add them to the
\code{additional_attributes} argument. Attribute names must match names in
\code{\link[biomaRt:listAttributes]{biomaRt::listAttributes()}}.
}
\examples{
\dontrun{
# Access version 110 instead of the latest Ensembl, and retrieve gene
# description during the query
gc_content <- get_gc_content_biomart(
  gene_ids = c("ENSG00000215388", "ENSG00000253771", "ENSG00000269584"),
  ensembl_version = 110,
  additional_attributes = c("description")
)

# Optionally pass in a pre-made biomaRt object
mart <- biomaRt::useEnsembl(
  biomart = "ensembl",
  dataset = "hsapiens_gene_ensembl"
)
gc_content <- get_gc_content_biomart(
  gene_ids = c("ENSG00000215388", "ENSG00000253771", "ENSG00000269584"),
  mart = mart
)
}
}
\seealso{
\link{get_gc_content_gtf}, \link{calculate_gc_content}
}
