% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{is_outlier_IQR}
\alias{is_outlier_IQR}
\title{Find Outliers by the IQR Method}
\usage{
is_outlier_IQR(data, tail = "both", IQR_mult = 1.5)
}
\arguments{
\item{data}{a numeric vector of data points}

\item{tail}{(optional) either "both", "lower", or "upper". If "lower", only
points below \code{Q1 - IQR_mult * IQR} will be marked as outliers. If "upper",
only points above \code{Q3 + IQR_mult * IQR} will be marked as outliers. If
"both", points in both tails will be marked as outliers. Defaults to
"both".}

\item{IQR_mult}{(optional) a multiplier to apply to the IQR. Defaults to 1.5.}
}
\value{
a vector of TRUE/FALSE values, where TRUE indicates that the point is
an outlier. Any \code{NA} values in \code{data} will be marked as TRUE.
}
\description{
Labels points as outliers if they are below \code{Q1 - 1.5 * IQR} and/or above
\code{Q3 + 1.5 * IQR}, with the option to modify the multiplier.
}
\details{
Note that \code{NA} values in \code{data} are ignored in the calculation of \code{Q1}, \code{Q3},
and \code{IQR}, and all \code{NA} data points will be marked as outliers in the final
result.
}
\examples{
data <- rnorm(n = 1000)
outliers <- is_outlier_IQR(data = data, tail = "upper")
print(data[outliers])
}
\seealso{
\code{\link[=is_outlier_SD]{is_outlier_SD()}}
}
