% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{is_outlier_SD}
\alias{is_outlier_SD}
\title{Find Outliers using Mean and Standard Deviation}
\usage{
is_outlier_SD(data, tail = "both", n_sds = 4)
}
\arguments{
\item{data}{a numeric vector of data points}

\item{tail}{(optional) either "both", "lower", or "upper". If "lower", only
points below \code{mean - n_sds * SD} will be marked as outliers. If "upper",
only points above \code{mean + n_sds * SD} will be marked as outliers. If
"both", points in both tails will be marked as outliers. Defaults to
"both".}

\item{n_sds}{(optional) the number of standard deviations away from the mean
that a point needs to be in order to be marked as an outlier. Defaults to
4.}
}
\value{
a vector of TRUE/FALSE values, where TRUE indicates that the point is
an outlier. Any \code{NA} values in \code{data} will be marked as TRUE.
}
\description{
Labels points as outliers if they are below \code{mean - n_sds * sd} and/or above
\code{mean + n_sds * sd}, where \code{n_sds} is a configurable multiplier.
}
\details{
Note that \code{NA} values in \code{data} are ignored in the calculation of \code{mean} and
\code{sd}, and all \code{NA} data points will be marked as outliers in the final
result.
}
\examples{
data <- rnorm(n = 1000)
outliers <- is_outlier_SD(data = data, n_sds = 3)
print(data[outliers])
}
\seealso{
\code{\link[=is_outlier_IQR]{is_outlier_IQR()}}
}
