% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{load_multiqc_json}
\alias{load_multiqc_json}
\title{Load MultiQC JSON file}
\usage{
load_multiqc_json(files)
}
\arguments{
\item{files}{a single path to a file or a vector or list of paths, which can
be relative or absolute paths}
}
\value{
a data.frame with fields from the following metrics: picard, rsem,
samtools, samtools idx, cutadapt. If multiple JSON files were supplied, all
data will be concatenated into one data.frame.
}
\description{
Loads stats from a MultiQC JSON file and merges them all into one data frame.
}
\details{
NOTE: This function has not been thoroughly tested.
}
\examples{
\dontrun{
# One file
stats_df <- load_multiqc_json("data/multiqc_data.json")

# Multiple files
stats_df <- load_multiqc_json(
  c("data/multiqc_data_1.json", "data/multiqc_data_2.json", "data/multiqc_data_3.json")
)
}
}
