% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{plot_sex_mismatch_results}
\alias{plot_sex_mismatch_results}
\title{Plot Sex Mismatch Results}
\usage{
plot_sex_mismatch_results(sex_check_df, y_expr_threshold = 2)
}
\arguments{
\item{sex_check_df}{a data.frame containing \code{reported_sex}, \code{estimated_sex},
and expression values for \code{XIST} and \code{mean_Y}, as output by
\code{\link[=find_sex_mismatches]{find_sex_mismatches()}}. It should also contain a \code{sex_valid} column
denoting whether the reported sex matches the estimated sex.}

\item{y_expr_threshold}{(optional) a numeric threshold that splits samples
based on mean expression of the Y-chromosome marker genes. The threshold
will be drawn as a line across the plots. Defaults to 2.0.}
}
\value{
a list of two \link[ggplot2:ggplot]{ggplot2::ggplot} plots, where the first plot is colored by
reported sex and the second plot is colored by mismatch status. Each plot
is also printed out.
}
\description{
This function plots the expression of XIST vs the mean expression of four
Y-chromosome genes (RPS4Y1, EIF1AY, DDX3Y, KDM5D) for each sample and
highlights mismatches between reported and estimated sex.
}
\examples{
\dontrun{
test
}
}
\seealso{
\code{\link[=find_sex_mismatches]{find_sex_mismatches()}}
}
