% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{remove_correlated_covariates}
\alias{remove_correlated_covariates}
\title{Remove Highly Correlated Covariates}
\usage{
remove_correlated_covariates(
  data,
  id_cols = c(),
  mixed_effects = c(),
  always_keep = c(),
  R2_threshold = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{data}{a data.frame or data.frame-like object of covariate data where
rows are samples and columns are covariates}

\item{id_cols}{(optional) a string or character vector of columns that should
be treated like IDs and never compared to other covariates or considered
for removal. Defaults to an empty vector.}

\item{mixed_effects}{(optional) a string or character vector of columns that
will be mixed effects in a regression model. These variables are compared
to other categoricals for potential removal but are not compared with
numeric variables, due to fact that mixed variables try to model any
unwanted clustering of numeric values by group. This effect might result in
high correlation between the mixed effect and numeric variable, but we
don't actually want to remove one of those variables if that's the case.
Defaults to an empty vector.}

\item{always_keep}{(optional) a string or character vector of columns that
should always be kept even if they are highly correlated with another
variable. In that case, the other variable is removed instead, as long as
it is not also in \code{always_keep}. Defaults to an empty vector.}

\item{R2_threshold}{(optional) the threshold for R^2, where variables with an
R^2 value > \code{R2_threshold} are considered for removal. Defaults to 0.5.}

\item{verbose}{(optional) whether to print out which variables have been
removed before returning. Defaults to \code{TRUE}.}
}
\value{
a data.frame that is a copy of \code{data} but without columns that meet
the removal criteria
}
\description{
This function examines the correlation between all given covariates and
removes at least one covariate from all pairs of highly-correlated covariates
in order to reduce the total correlation between variables in a regression
formula. \code{\link[=remove_unusable_covariates]{remove_unusable_covariates()}} should be run before using this
function, to remove variables that are all \code{NA} or all the same value.
}
\examples{
\dontrun{
}
}
