% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{remove_unusable_covariates}
\alias{remove_unusable_covariates}
\title{Remove Covariates that are Not Usable for Regression}
\usage{
remove_unusable_covariates(data, always_keep = c(), verbose = TRUE)
}
\arguments{
\item{data}{a data.frame or data.frame-like object of covariate data where
rows are samples and columns are covariates}

\item{always_keep}{(optional) a single string or a character vector of
variables to never remove even if they meet the criteria for removal, for
example a column of IDs. Defaults to an empty vector (removes everything
that meets the criteria).}

\item{verbose}{(optional) whether to print out which variables have been
removed before returning. Defaults to \code{TRUE}.}
}
\value{
a data.frame that is a copy of \code{data} but without columns that meet
the removal criteria
}
\description{
Given a data frame of metadata, this function removes columns that are either
all \code{NA} values, all the same value, or are not numeric and have a unique
value for each sample. Columns that have all the same value after removal of
\code{NA} values are also removed.
}
\examples{
df <- data.frame(
  id = 1:10,
  valid = rep(TRUE, 10),
  group = c(rep(1, 5), rep(2, 5))
)
df <- remove_unusable_covariates(df, always_keep = c("id"))
}
