% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_functions.R
\name{simple_lognorm}
\alias{simple_lognorm}
\title{Simple log2-Normalization of a Counts Matrix}
\usage{
simple_lognorm(data, pseudocount = 0.5)
}
\arguments{
\item{data}{a matrix or matrix-like object where rows are genes and columns
are samples. Data should be un-normalized counts. All values in \code{data}
should be >= 0.}

\item{pseudocount}{a pseudocount to add to every CPM value when taking the
log2, to avoid taking log2(0). Defaults to 0.5}
}
\value{
a matrix the same shape as data with all counts transformed to
log2(CPM)
}
\description{
Normalizes a counts matrix by transforming to CPM values and taking
log2(cpm + pseudocount)
}
\examples{
# Example using a pseudocount of 1, which makes all log2 values positive and
# count values of 0 remain 0.
counts <- round(matrix(runif(1000, min = 0, max = 1e8), nrow = 100))
log_data <- simple_lognorm(counts, pseudocount = 1)
}
\seealso{
\code{\link[=simple_cpm]{simple_cpm()}}
}
