% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{string_to_seed}
\alias{string_to_seed}
\title{Convert String to Numerical Seed}
\usage{
string_to_seed(string)
}
\arguments{
\item{string}{a string or a character vector or list. If the input is a
character vector/list, a seed will be generated for each item in the vector
separately}
}
\value{
an integer, if \code{string} is a single string, or a vector of integers
the same length as \code{string}, if \code{string} is a vector or list
}
\description{
Converts a string or character vector to an integer that can be used to set a
random seed.
}
\details{
This function is especially useful for things like running a process under
different conditions and ensuring that each run/condition gets its own unique
but consistent seed.
}
\examples{
# Single string
seed <- string_to_seed("Step 1")
set.seed(seed)

# Vector of strings
seeds <- string_to_seed(c("Step 1", "Step 2", "Step 3"))
set.seed(seeds[1])
}
