% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{find_pca_outliers}
\alias{find_pca_outliers}
\title{Find Outliers by PCA}
\usage{
find_pca_outliers(
  data,
  n_sds = 4,
  metadata = NULL,
  sample_colname = "specimenID",
  gene_info = NULL
)
}
\arguments{
\item{data}{a matrix, data.frame, or matrix-like object of expression data
where rows are genes and columns are samples. Data should be normalized and
on the log or log2 scale, for example as returned by \code{\link[=simple_lognorm]{simple_lognorm()}}.}

\item{n_sds}{(optional) samples will be labeled as outliers if they are
outside the ellipse defined by \code{radius1 = n_sds * sd(PC1)} and \code{radius2 = n_sds * sd(PC2)}. Defaults to 4.}

\item{metadata}{(optional) a data.frame where rows are samples and columns
are covariates. If supplied, it must contain, at minimum, a column with
sample IDs, and any additional columns will be merged into the result to
make plotting simpler. If omitted or \code{NULL}, the result will contain only
the columns returned in the \verb{$x} matrix from \code{prcomp}. See details for more
information.}

\item{sample_colname}{(optional) if \code{metadata} is provided, a character
string naming the column in \code{metadata} that contains sample labels. Ignored
if \code{metadata} is \code{NULL}. Defaults to "specimenID".}

\item{gene_info}{(optional) a data.frame where rows are genes and columns are
variables that provide metadata about each gene. If provided, the PCA will
run using only protein-coding autosomal genes, which are extracted from
\code{gene_info}. If \code{NULL} or not provided, the PCA will use all genes in
\code{data}. Defaults to \code{NULL}. See details for more information.}
}
\value{
a named list with the following items:

\item{pca_df}{the \verb{$x} data.frame returned by \code{prcomp}, plus any columns from
\code{metadata} if it was supplied}
\item{pc1_threshold, pc2_threshold}{the calculated radii of the ellipse for
PC1 and PC2}
\item{outliers}{a character vector of sample names that were marked as outliers}

The first three items in the list can be used to plot the results with
\code{\link[=plot_pca_outliers]{plot_pca_outliers()}}.
}
\description{
Runs PCA on normalized data and detects outliers that are a certain number of
standard deviations away from the mean on either PC1 or PC2.
}
\details{
This function runs \code{prcomp} on the provided data and marks samples as
outliers if they fall outside an ellipse defined by \code{radius1 = n_sds * sd(PC1)} and \code{radius2 = n_sds * sd(PC2)}.

\strong{Gene filtering}

The PCA can run using either protein-coding autosomal genes only, or all
genes in the expression matrix. Either set of genes is then filtered to genes
that are expressed in at least 80\% of samples and have a variance > 0.001, to
remove genes that are all 0 or have extremely low variance.

To run using protein-coding genes, a \code{gene_info} data.frame must be supplied
that contains the following columns: \code{ensembl_gene_id}, \code{gene_biotype},
\code{chromosome_name}. The values in \code{ensembl_gene_id} should contain values in
the rownames of \code{data}, and all protein-coding genes should have a
\code{gene_biotype} value of "protein_coding". This data frame can be obtained
with \code{\link[=get_gc_content_biomart]{get_gc_content_biomart()}} or \code{\link[=get_gc_content_gtf]{get_gc_content_gtf()}}, or built with
other tools to match the required structure.

\strong{Plotting}

This function returns all the information required to easily plot the results
with \code{\link[=plot_pca_outliers]{plot_pca_outliers()}}. If \code{metadata} is provided, the result will also
include all columns that exist in \code{metadata}, enabling the use of those
columns with \link[ggplot2:aes]{ggplot2::aes} specifications like color and shape.
}
\examples{
\dontrun{
}
}
\seealso{
\code{\link[=find_pca_outliers_by_group]{find_pca_outliers_by_group()}}, \code{\link[=simple_lognorm]{simple_lognorm()}},
\code{\link[=plot_pca_outliers]{plot_pca_outliers()}}, \code{\link[=get_gc_content_biomart]{get_gc_content_biomart()}}, \code{\link[=get_gc_content_gtf]{get_gc_content_gtf()}}
}
