% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{plot_pca_outliers}
\alias{plot_pca_outliers}
\title{Plot PCA Outlier Detection Results}
\usage{
plot_pca_outliers(
  pca_df,
  pc1_threshold,
  pc2_threshold,
  print_plot = TRUE,
  color = "is_outlier",
  ...
)
}
\arguments{
\item{pca_df}{a data.frame where rows are samples and columns are PCs as
output by \code{\link[=prcomp]{prcomp()}}, plus any additional covariates added by
\code{\link[=find_pca_outliers]{find_pca_outliers()}}.}

\item{pc1_threshold}{the outlier threshold for PC1, which is calculated by
\code{\link[=find_pca_outliers]{find_pca_outliers()}} as \code{n_sds * sd(PC1)}.}

\item{pc2_threshold}{the outlier threshold for PC2, which is calculated by
\code{\link[=find_pca_outliers]{find_pca_outliers()}} as \code{n_sds * sd(PC2)}.}

\item{print_plot}{(optional) whether to print the plot out before returning,
or not. Defaults to TRUE.}

\item{color}{(optional) the name of the column in \code{pca_df} to use as the
"color" in the \code{\link[ggplot2:aes]{ggplot2::aes()}} specification. If \code{color} is \code{NULL} or is
not a valid column name, all points will be black instead. Defaults to
\code{is_outlier}.}

\item{...}{(optional) additional \link[ggplot2:aes]{ggplot2::aes} specifications to pass to the
plot, in the format <aes_spec> = "column_name_as_string". For example,
\code{shape = "diagnosis"} or \code{size = "quantity"}.}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot} object with the built plot.
}
\description{
Plots PC1 vs PC2 and marks outlier samples as found by \code{\link[=find_pca_outliers]{find_pca_outliers()}}.
It also draws the ellipse that was used to determine outliers.
}
\examples{
\dontrun{
}
}
