% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{plot_sex_mismatch_results}
\alias{plot_sex_mismatch_results}
\title{Plot Sex Mismatch Results}
\usage{
plot_sex_mismatch_results(sex_check_df, y_expr_threshold)
}
\arguments{
\item{sex_check_df}{a data.frame containing \code{reported_sex}, \code{estimated_sex},
and expression values for \code{XIST} and \code{mean_Y}, as output by
\code{\link[=find_sex_mismatches]{find_sex_mismatches()}}. It should also contain a \code{sex_valid} column
denoting whether the reported sex matches the estimated sex.}

\item{y_expr_threshold}{(optional) a numeric threshold that splits samples
based on mean expression of the Y-chromosome marker genes. Samples with
mean expression above the threshold will be labeled as "male" while samples
with mean expression below the threshold will be labeled as "female".
Defaults to 2.0, though this value may need to be be calibrated to the
dataset.}
}
\value{
a list of two \link[ggplot2:ggplot]{ggplot2::ggplot} plots, where the first plot is colored by
reported sex and the second plot is colored by mismatch status. Each plot
is also printed out.
}
\description{
This function plots the expression of XIST vs the mean expression of four
Y-chromosome genes (RPS4Y1, EIF1AY, DDX3Y, KDM5D) for each sample and
highlights mismatches between reported and estimated sex.
}
\examples{
\dontrun{
test
}
}
\seealso{
\code{\link[=find_sex_mismatches]{find_sex_mismatches()}}
}
