% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_functions.R
\name{simple_cpm}
\alias{simple_cpm}
\title{Simple CPM Operation on a Counts Matrix}
\usage{
simple_cpm(data)
}
\arguments{
\item{data}{a matrix or matrix-like object where rows are genes and columns
are samples. All values in \code{data} should be >= 0.}
}
\value{
an object the same type and shape as data with all counts transformed
to CPM
}
\description{
Transforms count data into counts per million (CPM) by dividing by library
size and multiplying by 1 million.
}
\details{
This function is named \code{simple_cpm} to avoid confusion with \code{\link[edgeR:cpm]{edgeR::cpm()}},
which handles pseudocounts differently and has the ability to incorporate
TMM values into the library size.
}
\examples{
counts <- round(matrix(runif(1000, min = 0, max = 1e8), nrow = 100))
cpm_data <- simple_cpm(counts)
all(colSums(cpm_data) == 1e6)
}
\seealso{
\code{\link[=simple_lognorm]{simple_lognorm()}}
}
